metrics: use VictoriaMetrics instead of prometheus/client_golang
Most notably:
* the syntax is much simpler (no Register, less verbose calls)
* now we can Get metrics (concurrent-safe, along with Set), so:
  * data/ipfs/metrics.go now uses Set and Get instead of atomic.Float64
  * vochaininfo also uses Set and Get for most metrics, out of vi.lock.Lock()

Additional changes:
* unify types of vochaininfo to uint64
* VochainInfo.Height now returns uint64
* drop metric vochain_vote_tree_increase_last_minute
* drop vs.MetricsAgent field
* drop package `go.vocdoni.io/dvote/metrics` altogether
  (NewAgent was refactored into a new method in HTTProuter.ExposePrometheusEndpoint)

+++ api/api_types.go
@@ -198,7 +198,7 @@ type GenericTransactionWithInfo struct {
 
 type ChainInfo struct {
 	ID                      string    `json:"chainId" example:"azeno"`
-	BlockTime               [5]int32  `json:"blockTime" example:"12000,11580,11000,11100,11100"`
+	BlockTime               [5]uint64 `json:"blockTime" example:"12000,11580,11000,11100,11100"`
 	ElectionCount           uint64    `json:"electionCount" example:"120"`
 	OrganizationCount       uint64    `json:"organizationCount" example:"20"`
 	GenesisTime             time.Time `json:"genesisTime"  format:"date-time" example:"2022-11-17T18:00:57.379551614Z"`
